requirements.txt
assemblyai==0.36.0
spacy==3.8.3
# Add the following line to ensure the model is downloaded
spacy[en_core_web_sm]
wordcloud==1.9.4
pylangacq==0.19.1
ffmpeg==1.4
openai-whisper==20240930
tiktoken==0.8.0
scikit-learn==1.6.1
pandas==2.2.3
openai==1.60.1
numpy==2.0.2
matplotlib==3.9.4
python-dotenv==1.0.1
rapidfuzz==3.12.1
xgboost==2.1.4

pyannote.audio 
librosa
torch
torchvision
torchaudio